updated for reduced complexity
