+++ GANDLF/data/augmentation/README.md
@@ -9,5 +9,5 @@
     - Ensure it returns a `torch.Tensor`.
     - If the new function is from a pre-defined package, put it under `GANDLF.data.augmentation.wrap_${package_name}.py`.
 - Wrap the new algorithm in a `_transform` function as shown in `GANDLF.data.augmentation.crop_external_zero_planes.crop_external_zero_planes_transform`. This is the algorithm's identifier for GaNDLF.
-- Add the algorithm's identifier to `GANDLF.data.augmentation.global.global_augs_dict` as appropriate.
+- Add the algorithm's identifier to `GANDLF.data.augmentation.__init__.global_augs_dict` as appropriate.
 - Call the new identified from the config using the `data_augmentation` key.
