+++ GANDLF/data/preprocessing/README.md
@@ -8,5 +8,5 @@
     - Add function in new submodule.
     - Wrap it in a `_transform` function using `torchio.Lambda` as shown in `GANDLF.data.preprocessing.threshold_and_clip.threshold_transform`.
     - The `_transform` function is the algorithm's identifier for GaNDLF.
-- Add the algorithm's identifier to `GANDLF.data.preprocessing.global.global_preprocessing_dict` as appropriate.
+- Add the algorithm's identifier to `GANDLF.data.preprocessing.__init__.global_preprocessing_dict` as appropriate.
 - Call the new identified from the config using the `data_preprocessing` key.
